pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bannerforge"
version = "0.2.0"
description = "Generate stylish banners from structured inputs (speaker, talk, date, venue, photo)."
readme = "README.md"
authors = [{ name = "PythonCDMX" }]
requires-python = ">=3.10"
dependencies = [
  "pillow>=10.3",
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "httpx>=0.27",
  "typer>=0.12",
  "python-dotenv>=1.0",
  "segno>=1.6.1"
]

[project.scripts]
bannerforge = "bannerforge.cli:app"